requirements.txt
python-dotenv==1.0.0
python-pptx==0.6.21
python-docx==0.8.11
google-generativeai==0.3.1
